requirements.txt
asgiref==3.7.2
bandit==1.7.7
black==22.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.4.1
Django==4.2.11
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.4.0
dparse==0.6.4
drf-yasg==1.21.7
exceptiongroup==1.2.0
Faker==23.2.1
flake8==4.0.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.2.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.6.1
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==21.0
pathspec==0.11.1
pbr==6.1.0
platformdirs==3.8.0
pluggy==1.4.0
psycopg2-binary==2.9.10
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.0.1
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-qt==4.4.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.9.4
robotframework==7.2
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safety==2.3.5
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.37
sqlparse==0.4.4
stevedore==5.4.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
